Inner Status Column + Subcalls on failed txs (#437)
* feat: add and populate inner status column

* refactor: move exit code fns to tools package

* feat: add inner status ok to genesis txs

* test: update tests

* fix: use correct status to update error metrics

* feat: add all exit codes

* feat: update var names and comments

* feat: InnerTxStatus -> SubcallStatus

* feat: InnerTxStatus -> SubcallStatus

* fix: change fn name to be more accurate

* test: update tests

* fix: rm redundant status check
